test: add and organize tests for custom component with FSMap
Refactor the test suite by splitting the monolithic test file into separate files for better maintainability and clarity. Each test file now focuses on a specific scenario: rendering a single component, rendering multiple components, and rejecting invalid main component paths.
